# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SynthCut API",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/projects/{project_id}/uploads": {
      "post": {
        "tags": [
          "uploads"
        ],
        "summary": "Create Upload",
        "description": "Open (or resume) a resumable multipart upload for one file.",
        "operationId": "create_upload_api_v1_projects__project_id__uploads_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadSessionOut"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "507": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Insufficient Storage"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/uploads/{session_id}": {
      "get": {
        "tags": [
          "uploads"
        ],
        "summary": "Get Upload",
        "description": "Upload state, including the parts storage already holds (for resume).",
        "operationId": "get_upload_api_v1_uploads__session_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadSessionOut"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "uploads"
        ],
        "summary": "Abort Upload",
        "operationId": "abort_upload_api_v1_uploads__session_id__delete",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetOut"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/uploads/{session_id}/parts": {
      "post": {
        "tags": [
          "uploads"
        ],
        "summary": "Sign Parts",
        "description": "Presigned PUT URLs for parts; each signature covers the part's Content-MD5.",
        "operationId": "sign_parts_api_v1_uploads__session_id__parts_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartSignRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartSignResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/uploads/{session_id}/progress": {
      "post": {
        "tags": [
          "uploads"
        ],
        "summary": "Report Progress",
        "operationId": "report_progress_api_v1_uploads__session_id__progress_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadProgressReport"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/uploads/{session_id}/complete": {
      "post": {
        "tags": [
          "uploads"
        ],
        "summary": "Complete Upload",
        "description": "Verify every part in storage, assemble the original and queue ingestion.",
        "operationId": "complete_upload_api_v1_uploads__session_id__complete_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetOut"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Gone"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiErrorBody": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ApiErrorBody"
      },
      "AssetKind": {
        "type": "string",
        "enum": [
          "video",
          "audio",
          "image",
          "other"
        ],
        "title": "AssetKind"
      },
...
      "AssetOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "kind": {
            "$ref": "#/components/schemas/AssetKind"
          },
          "status": {
            "$ref": "#/components/schemas/AssetStatus"
          },
          "original_filename": {
            "type": "string",
            "title": "Original Filename"
          },
          "extension": {
            "type": "string",
            "title": "Extension"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "sort_index": {
            "type": "integer",
            "title": "Sort Index"
          },
          "etag": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Etag"
          },
          "sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sha256"
          },
          "duration_sec": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Sec"
          },
          "width": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width"
          },
          "height": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          },
          "fps": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fps"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "uploaded_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uploaded At"
          },
          "ready_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ready At"
          },
          "upload": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UploadStateOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "kind",
          "status",
          "original_filename",
          "extension",
          "content_type",
          "size_bytes",
          "sort_index",
          "etag",
          "sha256",
          "duration_sec",
          "width",
          "height",
          "fps",
          "error",
          "created_at",
          "uploaded_at",
          "ready_at"
        ],
        "title": "AssetOut"
      },
      "AssetStatus": {
        "type": "string",
        "enum": [
          "uploading",
          "uploaded",
          "ingesting",
          "ready",
          "failed",
          "cancelled"
        ],
        "title": "AssetStatus"
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorBody"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PartSignItem": {
        "properties": {
          "number": {
            "type": "integer",
            "maximum": 10000.0,
            "minimum": 1.0,
            "title": "Number"
          },
          "md5_b64": {
            "type": "string",
            "pattern": "^[A-Za-z0-9+/]{22}==$",
            "title": "Md5 B64",
            "description": "base64 of the part's 16-byte MD5"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "number",
          "md5_b64"
        ],
        "title": "PartSignItem"
      },
      "PartSignRequest": {
        "properties": {
          "parts": {
            "items": {
              "$ref": "#/components/schemas/PartSignItem"
            },
            "type": "array",
            "maxItems": 50,
            "minItems": 1,
            "title": "Parts"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "parts"
        ],
        "title": "PartSignRequest"
      },
      "PartSignResponse": {
        "properties": {
          "parts": {
            "items": {
              "$ref": "#/components/schemas/SignedPart"
            },
            "type": "array",
            "title": "Parts"
          }
        },
        "type": "object",
        "required": [
          "parts"
        ],
        "title": "PartSignResponse"
      },
...
      "SignedPart": {
        "properties": {
          "number": {
            "type": "integer",
            "title": "Number"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "method": {
            "type": "string",
            "title": "Method",
            "default": "PUT"
          },
          "headers": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Headers"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "number",
          "url",
          "headers",
          "expires_at"
        ],
        "title": "SignedPart"
      },
...
      "UploadCreate": {
        "properties": {
          "filename": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Filename"
          },
          "size_bytes": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Size Bytes"
          },
          "content_type": {
            "type": "string",
            "maxLength": 127,
            "title": "Content Type",
            "default": "application/octet-stream"
          },
          "last_modified_ms": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Modified Ms"
          },
          "fingerprint": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Fingerprint"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "filename",
          "size_bytes",
          "fingerprint"
        ],
        "title": "UploadCreate"
      },
      "UploadPartOut": {
        "properties": {
          "number": {
            "type": "integer",
            "title": "Number"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "etag": {
            "type": "string",
            "title": "Etag"
          }
        },
        "type": "object",
        "required": [
          "number",
          "size",
          "etag"
        ],
        "title": "UploadPartOut"
      },
      "UploadProgressReport": {
        "properties": {
          "bytes_uploaded": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Bytes Uploaded"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "bytes_uploaded"
        ],
        "title": "UploadProgressReport"
      },
      "UploadSessionOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "asset_id": {
            "type": "string",
            "format": "uuid",
            "title": "Asset Id"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "status": {
            "$ref": "#/components/schemas/UploadSessionStatus"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "part_size": {
            "type": "integer",
            "title": "Part Size"
          },
          "part_count": {
            "type": "integer",
            "title": "Part Count"
          },
          "uploaded_parts": {
            "items": {
              "$ref": "#/components/schemas/UploadPartOut"
            },
            "type": "array",
            "title": "Uploaded Parts"
          },
          "bytes_uploaded": {
            "type": "integer",
            "title": "Bytes Uploaded"
          },
          "resumed": {
            "type": "boolean",
            "title": "Resumed",
            "default": false
          },
          "asset": {
            "$ref": "#/components/schemas/AssetOut"
          }
        },
        "type": "object",
        "required": [
          "id",
          "asset_id",
          "project_id",
          "status",
          "size_bytes",
          "part_size",
          "part_count",
          "uploaded_parts",
          "bytes_uploaded",
          "asset"
        ],
        "title": "UploadSessionOut"
      },
      "UploadSessionStatus": {
        "type": "string",
        "enum": [
          "active",
          "completing",
          "completed",
          "aborted",
          "expired",
          "failed"
        ],
        "title": "UploadSessionStatus"
      },
      "UploadStateOut": {
        "properties": {
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "status": {
            "$ref": "#/components/schemas/UploadSessionStatus"
          },
          "bytes_reported": {
            "type": "integer",
            "title": "Bytes Reported"
          },
          "part_size": {
            "type": "integer",
            "title": "Part Size"
          },
          "part_count": {
            "type": "integer",
            "title": "Part Count"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "status",
          "bytes_reported",
          "part_size",
          "part_count"
        ],
        "title": "UploadStateOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}